requirements.txt
aiogram == 2.23.1
flask
discord.py
requests
